backend/requirements.txt
fastapi
uvicorn
qdrant-client
numpy
requests
python-dotenv